pyproject.toml
[project]
name = "codeindex"
version = "0.2.0"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "CodeIndex Authors" }]
dependencies = [
    "openai[aiohttp]>=2.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "tree-sitter>=0.21.0,<0.22.0",
    "tree-sitter-languages>=1.10.2",
]


[project.scripts]
codeindex = "codeindex.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]

[tool.uv]
package = true
